+++ notifier-service/src/main/java/com/example/notifier/repository/TemplateRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TemplateRepository extends JpaRepository<Template, Integer> {
-    Template findTemplateById(int id);
+public interface TemplateRepository extends JpaRepository<Template, Long> {
+    Template findTemplateById(long id);
 }
